backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.6.0
dj-database-url==2.3.0
psycopg[binary]==3.2.4
python-dotenv==1.0.1
